Slightly simplify the clearing-of-expired-sessions feature
Replace the CLEAR_SESSIONS_PROBABILITY setting with
CLEAR_EXPIRED_SESSIONS_AFTER_N_REQUESTS so it's easy to understand at a
glance and easy to figure out what to set it to if the default of 100
isn't suitable. Also, this setting can be set to None to disable the
session-clearing feature altogether.