Jettison Bootstrap forms & template tags
In e6c3213, the django-bootstrap-form code was added to arcutils.
I think if we want to fork a project like that (because it's
unmaintained or for any other reason), then we should fork it in the
normal way (i.e., fork it an maintain it as a separate package) instead
of bundling it into arcutils.

Apart from that, I also don't think the use of generic template tags to
generate forms is all that useful. It causes headaches when they don't
do exactly what you want and you have to dig around to figure out what's
going on. Note: this is about providing a set of super-generic templates
and template tags; if a specific project has a need for this kind of
thing, that's a separate consideration.