Replace ChoiceEnum w/ Python 3 version
That is, ChoiceEnum now extends the built-in Enum type from Python 3.4.
A back-ported version of the enum module will be installed for Python
3.3.

The new version is fairly similar to the old version but doesn't require
any hackery since the functionality of the old version is natively
supported by the built-in Enum type (iteration, item access, &c).