Add masquerade app
This is similar to django-cloak but cleaned up and more configurable.
It allows staff and superusers (by default) to masquerade as other
users (except superusers).

I'm not sure if this should be moved into its own package. On the one
hand, that would allow other non-PSU/non-ARC people to use it, and it
would allow it to evolve separately from arcutils. On the other hand,
doing so would increase maintenance costs. Conundrum.