Clarify that ldap.parse_email() returns preferred email address
At least, that is what this change attempts to do. I'm not 100% sure
which LDAP email field contains the preferred address (there are three
separate email address fields).

By default `parse_email` gets the email address from the LDAP `mail`
field because I think that's the one that contains the user's preferred
address.

`parse_email` now has a new `field` keyword arg that can be specified as
`mailLocalAddress` to get the first alias (again, not 100% about this)
or as `mailRoutingAddress` to get the canonical address (again, ...).