In ldapsearch(), use registered LDAP connection if available
If a connection object isn't passed explicitly, look for a connection
object in the component registry before falling back to creating a
new connection automatically from settings.

This solves the problem of connections being created and opened for
every single LDAP search. A project can now construct a connection or
connection pool up front and stick it in the registry for `ldapsearch()`
to use.