Don't do LDAP searches in with block
Using the connection as a context manager causes it to automatically
bind and then unbind (i.e., "disconnect and close the connection"), but
we don't want to unbind after every search. I don't think so anyway; the
docs aren't very clear about this and the ldap3.Connection code isn't
very illuminating.