Remove logging module in favor of settings
Get rid of the `logging` module. It redefines Django's logging config
in a way that is complex and unnecessary.

Instead, add some additional logging config to the default settings. In
addition to default formatters and handlers, add the Logstash logging
config to the default production logging config.

In the current production setup, any INFO or higher log messages that
propagate to the root logger (noting that propagation is the default)
will be logged to both the console and to Logstash. What this means in
practice:

- Django error messages due to uncaught exceptions or security issues
  will *not* be logged to the console or Logstash (because the
  `django.request` and `django.security` loggers don't propagate)
- Django INFO+ messages *will* be logged
- App INFO+ messages *will* be logged
- INFO+ messages from other loggers that propagate *will* be logged

Potential problems with the abover are:

- We *may* want to send Django error messages to the console and
  Logstash
- We may *not* want to send INFO messages from Django or other packages
  to the console and Logstash

The current setup is *okay*, but we'll probably need to revisit this
later.