Fix documentation in wsgi module
We can't recommend importing `create_wsgi_application` from arcutils and
then calling it to create the WSGI `application` global. That won't work
in ARC production environments because `arcutils` isn't on `sys.path`
when the project's `wsgi` module is imported by mod_wsgi.

Instead, change the docs to say that the code in the `wsgi` module must
be *copied* into the project's `wsgi` module.