Allow `@cached_property`s to specify their dependencies
And add `CachedPropertyInvalidatorMixin` which handles the invalidation
of cached properties.

E.g.:

    class Type(CachedPropertyInvalidatorMixin):

       def __init__(self, thing):
           self.thing = thing

       @cached_property('thing')
       def prop(self):
           return something if self.thing else something_else

Now, if `self.thing` is set to some other value or deleted, `prop` will
be invalidate (i.e., it will be recomputed the next time it's accessed).

Note: If `CachedPropertyInvalidatorMixin` isn't mixed in, specifying
dependencies has no effect.