Reimplement Registry internals using Option types
`Registry._find_component()` now returns an `Option`. Methods that use
`_find_component()` have been updated accordingly.

This work also revealed that `_find_component()` needs to return the
actual key under which the component is found because it may differ from
the key that `_find_component()` was called with (because components can
be registered under a subtype and looked up using a supertype).

This probably seems pointless in a way, but the internals are quite a
bit simpler now.