Don't explicitly raise TypeErrors in Option methods
A couple Option methods used to raise a supposedly more-informative
TypeError when a non-callable was passed. Instead, let non-callables be
called and raise a TypeError as usual.

The other option would be to explicitly raise TypeErrors in *all* Option
methods when a non-callable is passed, but I don't see much value in
doing that.