Make LDAP functions support AD lookups
For who-knows-what-reason, some of the fields in AD have names that are
different from the corresponding fields in "regular" OIT LDAP.

These changes account for that by looking for AD fields when "regular"
fields aren't found. This should probably be done a different way that's
more precise (e.g., specify the lookup type).