Note in ldap module that we should use ldap3's abstraction layer
We're doing a bunch of manual parsing that it looks like ldap3 already
handles with its abstraction layer.

It's somewhat unfortunate that we weren't using that from the start
because retrofitting it is going to be a pain. That said, it will be
worth the effort, since it will mean less code for us to maintain, and
I'm almost certain the authors of ldap3 have done a better job parsing
LDAP results than we ever could.