Add CAS backend
This is a simplified version of https://github.com/kstateome/django-cas,
which claims to be maintained but hasn't been updated in a while.

This version is stripped down, removing a bunch of stuff that we don't
use.

It's also more flexible, in particular in that it's possible to
configure login and logout paths separately, which is something that
needed to be fixed in order to migrate to PSU's new CAS system (which
has a non-standard logout path).

This also adds a `CASModelBackend`, which is a hybrid of the built-in
`ModelBackend` and the `CASBackend`. It uses CAS to authenticate while
providing all the permissions handling from the `ModelBackend`. This is
something that will be needed in, e.g., QuickTicket 3.