Switch default LDAP host from ldap-bulk to ldap-login
I'm really not sure which one of these should be the default. It seems
nearly impossible to get definitive answers about such things, so this
is really just a guess.

Note that the immediate reason for making this change is that ldap-bulk
seems unstable in QuickTicket. I'm not sure if that's because ldap-bulk
is flaky or ARCUtils' LDAP connection settings need to be tweaked.

Since ldap-login was recommended by Elliot for having "more servers in
the pool", it seems like switching should be okay.

Note that Matt's documentation in the ARC Handbook recommended using
ldap-batch (probably meaning ldap-bulk) because it's "beefier" than
ldap-login, which is exactly the opposite of what Elliot said.

Also note that ldap-login does not actually require login.