Improve handling of registry component factories
In registry.Registry...

- when adding factories, wrap them in a custom type instead of adding
  them to a set
- atomically materialize component factories into components

In the previous implementation, one thread could remove a component
factory from the set and then get suspended. Then another thread could
come along and return the factory because it would look like the factory
had already been materialized (because it was no longer in the set).

The new implementation materializes the component inside a locked block
and ensures factories are materialized only once (since creating the
same component twice could be problematic in some cases).