Revert to doing LDAP searches in with block
This reverts 9cb2523. Apparently, not doing searches in a with block can
cause connection issues.

In particular, a connection can be opened, bound, and then used for X
amount of time or N searches (not entirely sure which). After a while,
the connection is terminated by the server, but ldap3 doesn't seem to
notice this. Thinking the connection is still open, ldap3 tries to reuse
it, which leads to a "socket is closed" type of error.

When doing searches in a with block, connections are opened and bound
on entry and then unbound on exit. When using a sync strategy, this
actually opens and closes the socket.