Remove an errant code path in registry.add_registry()
There was an untested path for when a registry is added with the same
name as an existing registry. If it had ever been invoked, it would have
blown up hard. It was found because we don't generally add multiple
registries, since we typically only use the default registry.

This must have been part of an initial prototype implementation that
didn't get removed.

Added a note to the add_registry docstring noting what happens if a
registry is added with the same name as an existing registry (the
existing registry will be replaced).