Rewrite db.will_be_deleted_with()
The notable thing is that it's possible to check if a Django ORM
instance is in a set simply by doing `instance in set_of_instances`.
There's no need to do a PK check. This is due to the way ORM instance
are hashed using their PK values.

There's also no need to construct a new set just to remove an item from
the original set: we can simply use `set.remove(instance)` instead.

Also: Clean up.