Revert to SYNC as default LDAP connection strategy
The RESTARTABLE strategy sounds like it would be a good default because
it's synchronous and seems like it would handle intermittent network
connectivity issues, but in practice it just ends up hammering the
server and eventually failing anyway.

The default should be simple, synchronous, and thread safe for the
common case.