Add ldap package
Split the code from the existing ldap module into various modules in
this new package and export a reasonable set of objects.

One backward-incompatible consequence of this is that the various
parse_* functions are no longer automatically exported. This shouldn't
affect any existing projects other than quickticket.

The next step will be to reimplement the profile-parsing functionality
to use classes so that parsing can be customized more easily (e.g., for
LDAP vs AD connections).