Back port cas package from ARCUtils v2
This is somewhat unfortunate, since it would be "better" to migrate
projects using ARCUtils v1 to v2, but we probably don't have time to do
that before the Shibboleth CAS deadline (3/28).

One thing that's different about this version is the `cas.settings`
package. We might want to actually forward port it (or the basic idea
anyway) to v2.