Allow overrides to be passed when creating users from CAS
CASBackend.create_user() now accepts keyword args that will be set on
the newly-created user. These args will override any user args pulled
from the CAS response and any other args set in create_user().

This allows for this:

    return super().create_user(cas_data, is_staff=True)

instead of this:

    user = super().create_user(cas_data)
    user.is_staff = True
    user.save()
    return user

The former is cleaner and avoids an additional database query.