Make require_block template tag more configurable
Previously, the prefix of the static path to the RequireJS script was
hard coded as 'vendor'. Now that prefix defaults to 'vendor' and there
is a setting that can be used to override it (e.g., to 'node_modules'):

    ARC.require_block.prefix

Historically, `require_block` assumed Bower was being used to install
front end libraries and that Bower was configured to install things in
`{static root}/vendor`. Now, on new projects and some existing projects,
we're using npm to install front end libraries.

This feels a little hacky. Hopefully we'll be moving away from using
RequireJS altogether soon in favor of ES6 modules.