Add redirect_location template tag
This is a wrapper around response.get_redirect_location().

It lets us do things like this:

    <a href="{% redirect_location default='some-route-name' %}">Cancel</a>
    <a href="{% redirect_location default='/some-url' %}">Cancel</a>

Instead of this:

    <a href="{% url some-route-name %}">Cancel</a>

The former may be preferable when it's better to send the user back to
where they came from instead of hard coding a location. Note that the
former also lets us set a default in case there's no referrer or the
referrer is from a different domain.