Add collectstatic command w/ better (in|ex)clude capabilities
This adds --exclude and --include options to the collectstatic command.
If these options aren't used, this custom command should behave the same
as the built in version.

--exclude is similar to the built in --ignore except that it matches
against entire (relative) paths instead of path segments.

--include matches against entire (relative) paths like --exclude. If a
file is included with --include, it will be included even if it was
previously ignored or excluded.

The basic idea is that you might want to exclude a big tree of stuff,
such as node_modules/, but then add back selected files, such as
almond.js.